split context
